OSIS-1700 Rename app
