+++ backoffice/settings/base.py
@@ -65,7 +65,7 @@ INSTALLED_APPS = (
     'ckeditor',
     'osis_common',
     'reference',
-    'RulesManagement',
+    'rules_management',
     'base',
     'statici18n',
     'rest_framework',
