+++ rules_management/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'rules_management.apps.RulesmanagementConfig'
