+++ rules_management/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
 # Register your models here.
-from RulesManagement.models import FieldReference, FieldReferenceAdmin
+from rules_management.models import FieldReference, FieldReferenceAdmin
 
 admin.site.register(FieldReference, FieldReferenceAdmin)
