+++ rules_management/fatories.py
@@ -23,10 +23,35 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from django.apps import AppConfig
-from django.utils.translation import gettext_lazy
+
+import factory
+from django.contrib.auth.models import Permission
+from django.contrib.contenttypes.models import ContentType
 
 
-class RulesmanagementConfig(AppConfig):
-    name = 'RulesManagement'
-    verbose_name = gettext_lazy("Rules Management")
+class PermissionFactory(factory.DjangoModelFactory):
+    class Meta:
+        model = Permission
+
+    name = factory.Faker('text', max_nb_chars=255)
+    codename = factory.Faker('text', max_nb_chars=100)
+    content_type = factory.Iterator(ContentType.objects.all())
+
+
+class FieldReferenceFactory(factory.DjangoModelFactory):
+    class Meta:
+        model = 'rules_management.FieldReference'
+
+    content_type = factory.Iterator(ContentType.objects.all())
+    field_name = "acronym"
+
+    @factory.post_generation
+    def permissions(self, create, extracted, **kwargs):
+        if not create:
+            # Simple build, do nothing.
+            return
+
+        if extracted:
+            # A list of permissions were passed in, use them
+            for permission in extracted:
+                self.permissions.add(permission)
