+++ rules_management/tests.py
@@ -28,8 +28,8 @@ from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ImproperlyConfigured
 from django.test import TestCase
 
-from RulesManagement.fatories import PermissionFactory, FieldReferenceFactory
-from RulesManagement.mixins import PermissionFieldMixin
+from rules_management.fatories import PermissionFactory, FieldReferenceFactory
+from rules_management.mixins import PermissionFieldMixin
 from base.tests.factories.user import UserFactory
 from reference.models.country import Country
